+++ packages/cockpit/CHANGELOG.md
@@ -1,5 +1,35 @@
 # adonis-cockpit
 
+## 0.0.1-alpha.5
+
+### Patch Changes
+
+- Make resource icons optional
+
+- Updated dependencies []:
+  - @adonis-cockpit/lucid-filter@0.0.1-alpha.5
+  - @adonis-cockpit/lucid-infer@0.0.1-alpha.5
+
+## 0.0.1-alpha.4
+
+### Patch Changes
+
+- fix: fix some issues with configuration
+
+- Updated dependencies []:
+  - @adonis-cockpit/lucid-filter@0.0.1-alpha.4
+  - @adonis-cockpit/lucid-infer@0.0.1-alpha.4
+
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- feat: move resources registration to config file with lazyloading support
+
+- Updated dependencies []:
+  - @adonis-cockpit/lucid-filter@0.0.1-alpha.3
+  - @adonis-cockpit/lucid-infer@0.0.1-alpha.3
+
 ## 0.0.1-alpha.2
 
 ### Patch Changes