+++ packages/lucid-infer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @adonis-cockpit/lucid-infer
 
+## 0.0.1-alpha.5
+
+### Patch Changes
+
+- Make resource icons optional
+
+## 0.0.1-alpha.4
+
+### Patch Changes
+
+- fix: fix some issues with configuration
+
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- feat: move resources registration to config file with lazyloading support
+
 ## 0.0.1-alpha.2
 
 ### Patch Changes