+++ packages/react/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @adonis-cockpit/react
 
+## 0.0.1-alpha.5
+
+### Patch Changes
+
+- Make resource icons optional
+
+- Updated dependencies []:
+  - adonis-cockpit@0.0.1-alpha.5
+
+## 0.0.1-alpha.4
+
+### Patch Changes
+
+- fix: fix some issues with configuration
+
+- Updated dependencies []:
+  - adonis-cockpit@0.0.1-alpha.4
+
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- feat: move resources registration to config file with lazyloading support
+
+- Updated dependencies []:
+  - adonis-cockpit@0.0.1-alpha.3
+
 ## 0.0.1-alpha.2
 
 ### Patch Changes